requirements.txt
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.5
DateTime==5.2
greenlet==3.0.0
iniconfig==2.0.0
Jinja2==3.1.2
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pwinput==1.0.3
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==7.4.2
pytest-html==4.0.2
pytest-metadata==3.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
rich==13.6.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
zope.interface==6.1
